Code optimization and deleteCategories function - Moved topic and category functions to separate file - Fixed 'response' variable errors from cloud codes
